+++ packages/fiori-app-sub-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fiori-app-sub-generator
 
+## 0.13.10
+
+### Patch Changes
+
+- @sap-ux/fiori-elements-writer@2.8.116
+- @sap-ux/fiori-freestyle-writer@2.5.86
+- @sap-ux/odata-service-inquirer@2.20.7
+
 ## 0.13.9
 
 ### Patch Changes