+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.8.116
+
+### Patch Changes
+
+- Updated dependencies [791e9b9]
+    - @sap-ux/ui5-test-writer@0.7.97
+
 ## 2.8.115
 
 ### Patch Changes