+++ packages/flp-config-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/flp-config-inquirer
 
+## 0.4.167
+
+### Patch Changes
+
+- Updated dependencies [68b5523]
+    - @sap-ux/adp-tooling@0.18.108
+
 ## 0.4.166
 
 ### Patch Changes